migrations/default/1_20240625033210_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `event` ADD `next_run` DATETIME(6) NOT NULL;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE `event` DROP COLUMN `next_run`;"""
